script: add git script to compile every commit in a range of commits
This is a git script that will iterate through every commit in a
specified range, and perform a configure and make.  The intention of
this script is not to act as a check of code correctness, but to see if
any commit breaks compilation of the tree.

The range passed in to the script is in the form of a normal git range;
the default is HEAD^!  (i.e. just the latest commit).

Options may be passed in for configure, as well as make.  The log output
filename and directory may also be optionally specified.

If everything compiled without error, then the exit code from the
program will be 0.