git-foreach: new script to execute a command over a range of checkouts.
This script will execute a specified command over a range of commits.

Some things to note with the current version:

1. The command is not passed the commit hash

2. A 'git checkout' is executed for each commit hash in the range.
   This allows you to run arbitrary commands on each commit.

3. Currently, if you supress error output for the command, then
   the script will also not exit on the first error.

Example usage:

I often use this to run checkpatch.pl over a series of commits.  As
such, that is the default command to run.  But if you were to specify
the command fully from the commandline, it would look like:

git-foreach -r origin/master.. -c "./scripts/checkpatch.pl <(git format-patch -s --stdout HEAD^\!)" -n

(We have to escape the '!' from the command line, but not if this is
 specified in .git/config)

The above foreach command will run the checkpatch.pl script on each
commit on the current local HEAD, from the origin master branch.  This
is useful for development branches, for instance.