git-backport-diff: Default to newest commit
It is more likely that a backport refers to a newer commit than to an
old one.  One reason not to do so is because a later commit may
reference the existing commit in its message.  To get around this, only
look at commits that actually have a matching subject and ignore their
message body.

As a side effect, this generally makes the script faster because the
git-log process is quit once the first valid match is found.