git-backport-diff: Kill git-log after match found
d0c8cbd7ac5bb has indeed made finding a match faster, but the downside
is that the git-log process continues to run in the background even when
we no longer consume its output.  This is a problem particularly for
large patch series, where git-backport-diff may thus spawn hundreds of
subprocesses.

We don't need the git-log process after we found a match, so make it a
real job instead of an anonymous subprocess, which allows us to
terminate it after we have found a match.

Reported-by: Thomas Huth <thuth@redhat.com>
Fixes: d0c8cbd7ac5bb78c46d17e4fc04e77b6d3962baf